Move ExceptionGenerator type to phase 2 types
