+++ packages/core/phase1/exceptions/HO100321.ts
@@ -3,7 +3,7 @@ import errorPaths from "../../lib/exceptions/errorPaths"
 import isDummyAsn from "../../lib/isDummyAsn"
 import type Exception from "../../types/Exception"
 import { isAsnFormatValid } from "../lib/isAsnValid"
-import type { ExceptionGenerator } from "../types/ExceptionGenerator"
+import type { ExceptionGenerator } from "../../types/ExceptionGenerator"
 
 const HO100321: ExceptionGenerator = (hearingOutcome) => {
   const recordableOnPNCindicator = hearingOutcome.AnnotatedHearingOutcome.HearingOutcome.Case.RecordableOnPNCindicator
