+++ packages/core/phase1/exceptions/HO100322.ts
@@ -3,7 +3,7 @@ import errorPaths from "../../lib/exceptions/errorPaths"
 import isCaseRecordable from "../../lib/isCaseRecordable"
 import type Exception from "../../types/Exception"
 import isAdjourned from "../lib/result/isAdjourned"
-import type { ExceptionGenerator } from "../types/ExceptionGenerator"
+import type { ExceptionGenerator } from "../../types/ExceptionGenerator"
 
 const HO100322: ExceptionGenerator = (hearingOutcome) => {
   if (!isCaseRecordable(hearingOutcome)) {
