+++ packages/core/phase1/exceptions/HO100323.ts
@@ -3,7 +3,7 @@ import errorPaths from "../../lib/exceptions/errorPaths"
 import isCaseRecordable from "../../lib/isCaseRecordable"
 import type Exception from "../../types/Exception"
 import findException from "../lib/findException"
-import type { ExceptionGenerator } from "../types/ExceptionGenerator"
+import type { ExceptionGenerator } from "../../types/ExceptionGenerator"
 
 const hasHO100322 = (exceptions: Exception[], offenceIndex: number, resultIndex: number): boolean =>
   !!findException(
