+++ packages/core/types/ExceptionGenerator.ts
@@ -1,5 +1,5 @@
-import type { AnnotatedHearingOutcome } from "../../types/AnnotatedHearingOutcome"
-import type Exception from "../../types/Exception"
+import type { AnnotatedHearingOutcome } from "./AnnotatedHearingOutcome"
+import type Exception from "./Exception"
 
 export type ExceptionGenerator = (
   hearingOutcome: AnnotatedHearingOutcome,
